local-manifest: add the possibility to refresh a local manifest
This adds the possibility to play a downloaded content until it is completely downloaded. This approach has a little downside: it needs the local manifest to never be mutated and to be re-created from scratch each time we ask for it. This could be tedious for the user and might put pressure on the browser's GC. We will see how it goes.